Moving NoDataForLine exception class in the error package
