+++ seep-java2sdg/src/main/java/uk/ac/imperial/lsds/java2sdg/analysis/TEBoundaryAnalysis.java
@@ -37,6 +37,7 @@ import uk.ac.imperial.lsds.java2sdg.bricks.SDGAnnotation;
 import uk.ac.imperial.lsds.java2sdg.bricks2.TaskElementNature;
 import uk.ac.imperial.lsds.java2sdg.bricks2.Variable;
 import uk.ac.imperial.lsds.java2sdg.bricks2.TaskElement.TaskElementBuilder;
+import uk.ac.imperial.lsds.java2sdg.errors.NoDataForLine;
 import uk.ac.imperial.lsds.java2sdg.input.SourceCodeHandler;
 
 public class TEBoundaryAnalysis {
